Move 'getMainOfferQuote' function from /DetailsModal/utils.ts to /OfferNew/utils.ts. Fix import in /DetailsModal/index.tsx.
